feat: refuse unprovisionable tools at authoring time
