+++ crates/lns-artifact/Cargo.toml
@@ -17,6 +17,7 @@ oci-client = "0.16"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = "0.10"
+thiserror = "2"
 hex = "0.4"
 tar = "0.4"
